Identity added

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -3,6 +3,9 @@ using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
+using Contracts;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -10,7 +13,8 @@ namespace AuctionService.Controllers;
 
 [ApiController]
 [Route("api/auctions")]
-public class AuctionController(AuctionDbContext context, IMapper mapper) : ControllerBase
+public class AuctionController(AuctionDbContext context, IMapper mapper, IPublishEndpoint publishEndpoint)
+    : ControllerBase
 {
     [HttpGet]
     public async Task<ActionResult<IEnumerable<AuctionDto>>> GetAllAuctions(string? date)
@@ -40,6 +44,7 @@ public class AuctionController(AuctionDbContext context, IMapper mapper) : Contr
         return mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
@@ -47,16 +52,21 @@ public class AuctionController(AuctionDbContext context, IMapper mapper) : Contr
 
         // Todo: add current user as seller
 
-        auction.Seller = "test";
+        auction.Seller = User.Identity?.Name;
 
         await context.Auctions.AddAsync(auction);
+
+        var newAuction = mapper.Map<AuctionDto>(auction);
+
+        await publishEndpoint.Publish(mapper.Map<AuctionCreated>(newAuction));
 
         if (await context.SaveChangesAsync() == 0)
             return BadRequest("Couldn't create auction");
 
-        return CreatedAtAction(nameof(GetAuctionById), new { id = auction.Id }, mapper.Map<AuctionDto>(auction));
+        return CreatedAtAction(nameof(GetAuctionById), new { id = auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id:guid}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -67,13 +77,20 @@ public class AuctionController(AuctionDbContext context, IMapper mapper) : Contr
             return NotFound();
         }
 
-        // Todo: seller name == auction
+        if (auction.Seller != User.Identity?.Name)
+        {
+            return Forbid();
+        }
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
         auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
         auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
         auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
+
+        var actionToPublish = mapper.Map<AuctionUpdated>(updateAuctionDto);
+        actionToPublish.Id = auction.Id.ToString();
+        await publishEndpoint.Publish(actionToPublish);
 
         if (await context.SaveChangesAsync() == 0)
         {
@@ -83,6 +100,7 @@ public class AuctionController(AuctionDbContext context, IMapper mapper) : Contr
         return NoContent();
     }
 
+    [Authorize]
     [HttpDelete("{id:guid}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -91,9 +109,14 @@ public class AuctionController(AuctionDbContext context, IMapper mapper) : Contr
         if (auction == null)
             return NotFound();
 
-        // Todo: seller name == auction
-
+        if (auction.Seller != User.Identity?.Name)
+        {
+            return Forbid();
+        }
+        
         context.Auctions.Remove(auction);
+
+        await publishEndpoint.Publish(new AuctionDeleted { Id = auction.Id.ToString() });
 
         if (await context.SaveChangesAsync() == 0)
         {
